Move resource sizing and admission into a jobs/resources package
jobs/commands is 56 files and ~15,800 lines implementing 32 command
types. Resource policy -- how large each container may be, and who
currently holds the host's memory -- is a different concern that had
accreted into it.

The split exports exactly seven symbols: the four resolvers used by the
command files, plus JobMemoryBytes, TryAcquireMemory and
MemoryBudgetBytes used by the dispatcher. Everything else stays private:
the sizing constants, memoryBudget, clampMemory, the env-override
readers, and the admission counter itself.

ResolveContainerLimits and ResolveBuildLimits move here from
run_agent_step.go and build_static_site.go along with the env var names
they read. That is not tidying -- JobMemoryBytes calls both, so leaving
them in commands while commands imported resources would have been an
import cycle. Their tests move with them.

The result is a clean three-layer shape: hostinfo reports facts about
the machine, resources turns those into policy, and commands and
entrypoints consume it. That ordering is what prevents a recurrence of
the cycle already hit once in this change set, where host detection
could not live in jobs/commands because jobs/commands imports client and
the ping client needed the same detection.

No behaviour change: the diff is file moves plus identifier renames, and
the test count is unchanged at 99 before and after.
