+++ jobs/commands/run_assistant_session.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/deployment-io/deployment-runner/jobs/resources"
 	"io"
 	"os"
 	"path/filepath"
@@ -222,7 +223,7 @@ func (rs *RunAssistantSession) runSession(orgID, jobID, imageRef, workDirHost st
 	// resolveSessionLimits. Without this they inherited the Task-Step cap
 	// (the whole host budget on a small runner) and held it for the
 	// conversation's lifetime, blocking every deploy behind them.
-	sessionMemoryBytes := resolveSessionMemoryBytes()
+	sessionMemoryBytes := resources.ResolveSessionMemoryBytes()
 	containerID, err := createAgentboxContainer(dockerCtx, cli, agentboxSpawnSpec{
 		imageRef: imageRef, workDirHost: workDirHost, env: envVars, memoryBytes: sessionMemoryBytes,
 	})
